update host code and renames
