--- packages/host/README.md
@@ -1,1 +1,0 @@
-# container
