+++ packages/host/package.json
@@ -1,37 +1,13 @@
 {
-  "name": "host",
-  "version": "0.0.0",
-  "private": true,
-  "author": "WebAudioModules",
-  "license": "UNLICENSED",
-  "scripts": {
-    "clean": "del-cli dist .cache",
-    "start": "node server.js"
-  },
-  "dependencies": {
-    "chalk": "^4.1.1",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "parcel-bundler": "^1.12.5",
-    "pingpongdelay": "^0.0.0",
-    "quadrafuzz": "^0.0.0",
-    "sdk": "^0.0.0",
-    "synth101": "^0.0.0"
-  },
-  "browserslist": [
-    "last 2 Chrome versions",
-    "last 2 Firefox versions",
-    "last 2 Safari versions",
-    "not dead"
-  ],
-  "webaudiomodules": {
-    "livegain": "dist",
-    "pingpongdelay": "dist",
-    "quadrafuzz": "dist",
-    "synth101": "dist",
-    "wamexample": "dist"
-  },
-  "devDependencies": {
-    "del-cli": "^4.0.0"
-  }
+    "name": "host",
+    "version": "0.0.0",
+    "private": true,
+    "author": "WebAudioModules",
+    "license": "MIT",
+    "scripts": {
+      "start": "node server.js"
+    },
+    "dependencies": {
+      "express": "^4.17.2"
+    }
 }
