+++ packages/host/server.js
@@ -1,14 +1,6 @@
-const Bundler = require('parcel-bundler');
 const express = require('express');
-// const cors = require('cors');
 
 const app = express();
-
-// fix : https://github.com/parcel-bundler/parcel/issues/1315#issuecomment-523524186
-app.get('/', (req, res, next) => {
-	req.url = '/index.html';
-	next();
-});
 
 app.use((req, res, next) => {
 	res.header('Cross-Origin-Embedder-Policy', 'require-corp');
@@ -16,11 +8,8 @@ app.use((req, res, next) => {
 	next();
 });
 
-app.use('/packages', express.static('../'));
-
-const bundler = new Bundler('src/*.html');
-app.use(bundler.middleware());
+app.use('/', express.static('../'));
 
 app.listen(1234);
 
-console.info('Host server started on http://localhost:1234');
+console.info('Host server started on http://localhost:1234/host/');
