--- packages/host/src/index.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>Wasabi Host</title>
-	<!--
-	<script
-		src="https://wasabi.i3s.unice.fr/WebAudioPluginBank/bower_components/webaudio-controls2/webaudio-controls.js">
-	</script>
-	-->
-	<script
-		src="./bpf.js">
-	</script>
-</head>
-
-<body>
-	<header>
-		<h1 id="title">Example WAM Host</h1>
-	</header>
-	<main>
-		<p><b style="color:white">NEW: </b><a href="../../api/docs/typedoc/index.html">See WAM 2.0 API docs</a></p>
-		<p><b style="color:white">NEW: </b><a href="../../sdk-parammgr/docs/typedoc/index.html">See WAM 2.0 Parameter Manager docs</a></p>
-		<div id="wrapper">
-			<div id="left">
-				<audio id="player" src="https://wasabi.i3s.unice.fr/WebAudioPluginBank/BasketCaseGreendayriffDI.mp3"
-					controls loop crossOrigin="anonymous">
-				</audio>
-				<p>Select MIDI input device
-					<select id="selectMidiInput" class="widget">
-					</select>
-				</p>
-				<p>Select live input device
-					<select id="selectAudioInput" class="widget">
-					</select>
-				</p>
-
-				<p>
-					<button id="toggleLiveInput" class="widget">Live input:
-						<span style='color:#cc7c6e;'>NOT ACTIVATED</span>, click to toggle on/off!</button>
-				</p>
-				<form id="form">
-					<div id="left">
-						<div>
-							<p>
-								<label>Enter any WAM Plugin URL</label>
-								<input type="text" class="widget" name="pluginUrl" placeholder="Enter a plugin URL" />
-							</p>
-						</div>
-						<div>
-							<input type="submit" class="widget" value="LOAD PLUGIN" />
-						</div>
-						<p></p>
-						<div>
-							<ul id="midi-emitters">
-								<p>Midi</p>
-								<!--<li data-plugin-url="midiKeyboard/src">Midi Virtual keyboard</li>-->
-								<li data-plugin-url="simpleMidiKeyboard">Simple MIDI Keyboard</li>
-								<li data-plugin-url="VirtualMidiKeyboardNoSound/src">Midi Virtual keyboard No Sound (just emits events)</li>
-								<li data-plugin-url="randomnote">Random note generator</li>
-							</ul>
-							<ul id="examples">
-								<p>Pure JavaScript</p>
-								<li data-plugin-url="quadrafuzz/dist">Quadrafuzz</li>
-								<li data-plugin-url="quadrafuzz_without_builder/src">Quadrafuzz without builder</li>
-								<li data-plugin-url="disto_machine/src">Disto Machine without builder</li>
-								<li data-plugin-url="pingpongdelay/dist">PingPongDelay</li>
-								<li data-plugin-url="graphicEqualizer/src">Graphic Equaliser (pureJS)</li>
-								<li data-plugin-url="WamExample/src">WAM Example (pureJS)</li>
-								<li data-plugin-url="tinySynth/src">TinySynth, a GM syth without GUI. Accepts program changes, volume change, notes...</li>
-								<p>Faust</p>
-								<li data-plugin-url="faustPingPongDelay/plugin">faustPingPongDelay</li>
-								<li data-plugin-url="faustPingPongDelayDefaultUI">faustPingPongDelayDefaultUI</li>
-								<li data-plugin-url="GuitarAmpSim60s">Guitar AmpSim 60s (generated by faust IDE)</li>
-								<li data-plugin-url="StonePhaserStereo">Stone Phaser (generated by faust IDE)</li>
-								<li data-plugin-url="TS9_OverdriveFaustGenerated">TS9 Overdrive (generated by faust IDE)</li>
-								<li data-plugin-url="BigMuff">BigMuff Fuzz (generated by faust IDE)</li>
-								<li data-plugin-url="faustFlute">faust Flute MIDI</li>
-								<p>TypesScript</p>
-								<li data-plugin-url="livegain/dist/livegain">LiveGain</li>
-								<li data-plugin-url="livegain/dist/oscilloscope">Oscilloscope</li>
-								<li data-plugin-url="livegain/dist/spectroscope">Spectroscope</li>
-								<li data-plugin-url="livegain/dist/spectrogram">Spectrogram</li>
-								<p>CSound</p>
-								<li data-plugin-url="csoundPitchShifter/dist">Csound PitchShifter</li>
-								<p>PedalBoard</p>
-								<li data-plugin-url="pedalboard/dist">Pedalboard plugin</li>
-								<p>Instruments</p>
-								<li data-plugin-url="synth101/dist">Synth101</li>
-								<li data-plugin-url="obxd">OBXD</li>
-								<p>Templates</p>
-								<li data-plugin-url="WamExampleTemplate/src">WAM Example Template (pureJS)</li>
-								<li data-plugin-url="_pmjs/src">Template1 (ParamMgr + JS)</li>
-							</ul>
-						</div>
-
-				</form>
-			</div>
-			<div id="right">
-				<div id="midiKeyboard"></div>
-				<div id="pluginContainer">
-					<div id="mount"></div>
-					<br>
-					<button id="saveStateButton" class="widget" disabled>SAVE STATE</button>&nbsp;
-					<button id="restoreStateButton" class="widget" disabled>RESTORE STATE</button>
-				</div>
-				<div id="pluginInfoContainer">
-					<div id="pluginInfoDiv"></div>
-				</div>
-				<div id="pluginAutomationContainer">
-					<div id="pluginParamSelectorContainer">
-						<select name="pluginParamSelector" id="pluginParamSelector" class="widget">
-							<option value="-1" disabled selected>Add Automation...</option>
-						</select>
-						<label for="pluginAutomationLength">Duration (sec):</label>
-						<input name="pluginAutomationLength" id="pluginAutomationLength" type="number" class="widget" min="1" step="0.1" value="1" />
-						<input id="pluginAutomationApply" type="button" class="widget" value="Apply" />
-					</div>
-					<div id="pluginAutomationEditor">
-					</div>
-				</div>
-				<div id="sendMIDIContainer">
-					<button id="sendMIDINoteButton" class="widget">Send MIDI Note</button>
-				</div>
-			</div>
-		</div>
-	</main>
-
-	<script src="main.js"></script>
-</body>
-
-</html>
