+++ packages/template_midi_prammgr/src/descriptor.json
