+++ packages/template_prammgr/src/descriptor.json
